+++ lib/widgets/form_field_switch.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 getFormFieldDecoration({required final fieldName, required final bool isReadOnly}) {
   return InputDecoration(
     labelText: fieldName,
-    border: isReadOnly ? InputBorder.none : const OutlineInputBorder(),
+    border: isReadOnly ? InputBorder.none : null,
   );
 }