+++ lib/widgets/working.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class WorkingIndicator extends StatelessWidget {
+  const WorkingIndicator({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(child: SizedBox(height: 40, width: 40, child: CircularProgressIndicator()));
+  }
+}