+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Added
 
+- Use package [provider] for state management
 - Load from CSV [mymoney.mmcsv] file
 - Close and create a new file